requirements.txt
anthropic>=0.40.0
duckduckgo-search>=6.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
apscheduler>=3.10
pytz>=2024.1
